fix : 아이템 구매 단계 에서뒤로가기 feat : 관리자 아이템 삭제
뒤로가기가 출력 안하던 버그 수정
아이템 삭제 번호 입력시 없는 번호일 경우
해당 아이템 번호의 아이템이 존재 하지 않습니다.
츨

+++ finalShop_sinhyeongi/src/menu_admin/AdminItem.java
@@ -41,8 +41,11 @@ public class AdminItem implements MenuCommand  {
 			this.item.NewItem(item, category, price);
 		}
 		else if( sel == 2) {
-			System.out.println(item.GetLastItemNum());
-			int No = Util.getValue("► 삭제할 아이템 번호 입력"+item.GetLastItemNum()+"]", 1, item.GetLastItemNum());
+			if(item.GetLastItemNum() == 0) {
+				System.out.println("아이템 이 없습니다!");
+				return false;
+			}
+			int No = Util.getValue("► 삭제할 아이템 번호 입력 [ "+item.GetFistItemNum()+" -"+item.GetLastItemNum()+"]", item.GetFistItemNum(), item.GetLastItemNum());
 			cont.DeleteItem(No);
 		}
 		else if( sel == 3) {
